Added current semester detection
I do not yet have an example of an Excel file that contains data from the current semester, but if I did, I anticipate that the code in this commit is properly structured to function as follows:
If the top right table is wider than the bottom table's rightmost side, then the rightmost column of the top right table (both for 2200 and 3200) will be uploaded to the database.

Also, in the style of the "|| 0" assignments that I saw Daisy doing, I have added those to the frontend to further ensure that no null values are passed, even if they are null in the spreadsheet.