Fixed "continuous" and "total" logic
The logic for selecting continuous semesters will now get all semesters in the intermediate years, and only semesters at or after the start semester in the beginning year, and only semesters at or before the start semester in the ending year.
The logic for "total" now removes selected gender and ethnicity characteristics when the "total" option is selected under demographics.